# baseline/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/promotions/{code}": {
      "get": {
        "operationId": "AdminPromotionController_get",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPromotion"
        ]
      },
      "patch": {
        "operationId": "AdminPromotionController_update",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPromotion"
        ]
      },
      "delete": {
        "operationId": "AdminPromotionController_remove",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPromotion"
        ]
      }
    },
    "/api/admin/promotions/{code}/toggle": {
      "patch": {
        "operationId": "AdminPromotionController_toggle",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPromotion"
        ]
      }
    },
    "/api/admin/promotions/{code}/dry-run": {
      "post": {
        "operationId": "AdminPromotionController_dryRun",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPromotion"
        ]
      }
    },
    "/api/admin/promotions/{code}/locked-tenants": {
      "get": {
        "operationId": "AdminPromotionController_lockedTenants",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPromotion"
        ]
      }
    },
...
  },
  "info": {
    "title": "Edu Server API",
    "description": "教培 CRM 后端 API — Phase B.L3 contract source-of-truth",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}